Fix -Wextra 3/4.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@1201 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-crt/misc/mingw-fseek.c
@@ -56,7 +56,7 @@ __mingw_fwrite (const void *buffer, size_t size, size_t count, FILE *fp)
       printf ("__mingw_fwrite: current %I64u, actual %I64u\n", 
 	      current_position.QuadPart, actual_length.QuadPart);
 #endif /* DEBUG */
-      if (current_position.QuadPart > actual_length.QuadPart)
+      if ((size_t)current_position.QuadPart > (size_t)actual_length.QuadPart)
 	{
 	  static char __mingw_zeros[ZEROBLOCKSIZE];
 	  long long numleft;
